feat: wip 카드리스트 비즈니스 로직 분리 작업 중
customHook에서 필터링 전역 상태 받을 예정
