+++ src/types/meetingsType.ts
@@ -1,3 +1,17 @@
+export interface IMeeting {
+	teamId: number;
+	id: number;
+	type: string;
+	name: string;
+	dateTime: string;
+	registrationEnd: string;
+	location: string;
+	participantCount: number;
+	capacity: number;
+	image: string;
+	createdBy: number;
+	canceledAt: string;
+}
 export interface getMeetingParamsType {
 	type?: string;
 	location?: string;
@@ -5,4 +19,9 @@ export interface getMeetingParamsType {
 	sortOrder?: string;
 	limit?: number;
 	offset?: number;
+}
+
+export interface MeetingCardListProps {
+	initialData?: IMeeting[];
+	meetingType?: 'favorite' | undefined; // 찜한 목록
 }
